Add files via upload
The component fetching the details from the JSON is Pollution. Pollution component has two sister JS files wherein one JS file (pollutionActions) defines the actions for the redux store and the other (pollutionReducer) defines the logic for the reducer. The component is then rendered in the index.js. A search query to dynamically filter the data based on the city is also provided in the app for better functionality